dags/requirements.txt
backports.entry-points-selectable==1.1.0
boto3==1.18.60
botocore==1.21.60
cachetools==4.2.4
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.7
cycler==0.11.0
distlib==0.3.3
filelock==3.3.0
google-api-core==2.2.2
google-api-python-client==2.29.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
greenlet==1.1.2
httplib2==0.20.2
identify==2.3.0
idna==3.3
jmespath==0.10.0
kiwisolver==1.3.2
matplotlib==3.4.3
mysqlclient==2.0.3
nodeenv==1.6.0
numpy==1.21.4
oauth2client==4.1.3
oauthlib==3.1.1
Pillow==8.4.0
platformdirs==2.4.0
pre-commit==2.15.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydicom==2.2.2
pyparsing==3.0.5
python-dateutil==2.8.2
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.5.0
six==1.16.0
SQLAlchemy==1.4.25
toml==0.10.2
torch==1.10.0
torchaudio==0.10.0
torchvision==0.11.1
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
virtualenv==20.8.1